+++ filters/include/pcl/filters/impl/radius_outlier_removal.hpp
@@ -75,7 +75,7 @@ pcl::RadiusOutlierRemoval<PointT>::applyFilter (PointCloud &output)
   int nr_removed_p = 0;
   
   // Go over all the points and check which doesn't have enough neighbors
-  for (size_t cp = 0; cp < indices_->size (); ++cp)
+  for (int cp = 0; cp < static_cast<int>(indices_->size ()); ++cp)
   {
     int k = tree_->radiusSearch ((*indices_)[cp], search_radius_, nn_indices, nn_dists);
     // Check if the number of neighbors is larger than the user imposed limit
