+++ filters/src/extract_indices.cpp
@@ -78,7 +78,7 @@ pcl::ExtractIndices<sensor_msgs::PointCloud2>::applyFilter (PointCloud2 &output)
   {
     // Prepare a vector holding all indices
     std::vector<int> all_indices (input_->width * input_->height);
-    for (size_t i = 0; i < all_indices.size (); ++i)
+    for (int i = 0; i < static_cast<int>(all_indices.size ()); ++i)
       all_indices[i] = i;
 
     std::vector<int> indices = *indices_;
